RIS-3: add logging

+++ ris_3/repositories/column/model.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 from sqlalchemy import Column
 from clickhouse_sqlalchemy.engines import MergeTree
-from clickhouse_sqlalchemy.types import Int, String, Float, DateTime
+from clickhouse_sqlalchemy.types import Int, String, DateTime, Float64
 
 from ris_3.repositories.column.core import Base
 
@@ -18,7 +18,7 @@ class ColumnSale(Base):
 
     id = Column(Int, primary_key=True)
     name = Column(String(120), nullable=False)
-    price = Column(Float, nullable=False)
+    price = Column(Float64, nullable=False)
     shop = Column(String(120), nullable=False)
     sale_date = Column(
         DateTime, nullable=False, default=datetime.now
